Renaming and clear dead code
